+++ bardic/stdlib/__init__.py
@@ -2,6 +2,8 @@
 
 from .relationship import Relationship
 from .dice import roll, skill_check, weighted_choice, advantage, disadvantage
+from .inventory import Inventory
+from .economy import Wallet, Shop
 
 __all__ = [
     "Relationship",
@@ -10,4 +12,7 @@ __all__ = [
     "weighted_choice",
     "advantage",
     "disadvantage",
+    "Inventory",
+    "Wallet",
+    "Shop",
 ]